# docs/openapi.yaml
...
openapi: '3.1.0'
info:
  title: SomniBot Dashboard API
  version: '1.0.0'
  description: |
    Internal API for the SomniBot dashboard. All routes require guild-owner
    authentication via Discord OAuth session unless noted otherwise.

    **Authentication**: Discord OAuth2 session cookie (managed by NextAuth).
    Most endpoints also require `X-Guild-ID` header or derive guild from session.
  contact:
    email: heyimdionysus@gmail.com
servers:
  - url: '{dashboardUrl}'
    description: Dashboard instance
    variables:
      dashboardUrl:
        default: 'http://localhost:3000'
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie from Discord OAuth
...
paths:
...
  /api/customers/{id}/entitlements:
    get:
      summary: 'GET /api/customers/{id}/entitlements'
      responses:
        '200':
          description: Success
    post:
      summary: 'POST /api/customers/{id}/entitlements'
      description: 'Create an atomic zero-dollar non-commerce order and entitlement. Reuse request_id for every retry of the same grant. Omitted granted_role_ids/granted_channel_ids default to the product''s canonical sets; explicit lists are forwarded verbatim and must equal the canonical sets exactly.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [request_id, product_id]
              properties:
                request_id:
                  type: string
                  format: uuid
                  description: 'Caller-generated replay identity; reuse after a lost response.'
                product_id:
                  type: string
                  format: uuid
                type:
                  type: string
                  enum: [one_time, subscription]
                  default: one_time
                plan_id:
                  type: string
                  format: uuid
                  description: 'Required when type is subscription; forbidden for one_time.'
                source:
                  type: string
                  enum: [manual, giveaway, automation]
                  default: manual
                expires_at:
                  type: string
                  format: date-time
                granted_role_ids:
                  type: array
                  items:
                    type: string
                  description: 'Omit to grant the product''s canonical role set. If supplied, must equal it exactly.'
                granted_channel_ids:
                  type: array
                  items:
                    type: string
                  description: 'Omit to grant the product''s canonical channel set. If supplied, must equal it exactly.'
      responses:
        '200':
          description: 'Success — { id, order_id, request_id }'
        '400':
          description: 'Validation failed (bad JSON, missing product_id, malformed lists)'
        '404':
          description: 'Customer, product, or active subscription plan not found in this guild'
        '409':
          description: 'Grant conflicts with the product''s canonical access contract (RPC 23514)'
    put:
      summary: 'PUT /api/customers/{id}/entitlements'
      responses:
        '200':
          description: Success
...